Robot Arm demo sketch added
This is the Robot Arm demo sketch used for the video: https://www.youtube.com/watch?v=R0Ovw1ZrGzg
The code is provided as is. A description of the functionality is provided in the sketch. Please be aware that the Robot Arm may do moves that can damage itself or it surroundings. By using this code, you acknowledge that the authors can not be held responsible for damage.